suppress warnings

+++ src/ak_accessor.c
@@ -17,7 +17,7 @@ ak_accessor_setparams(AkAccessor     *acc,
   AkDataParam *dp, *last_dp;
   uint32_t     i;
 
-  assert(semantic > 0 && semantic < ak_def_semanticc());
+  assert(semantic > 0 && semantic < (int)ak_def_semanticc());
 
   smpair = ak_def_semantic()[semantic];
   acc->bound = smpair->count;
